+++ src/Jackalope/RepositoryFactoryJackrabbit.php
@@ -41,6 +41,7 @@ class RepositoryFactoryJackrabbit implements RepositoryFactoryInterface
         'jackalope.check_login_on_server' => 'boolean: if set to empty or false, skip initial check whether repository exists. Enabled by default, disable to gain a few milliseconds off each repository instantiation.',
         'jackalope.disable_stream_wrapper' => 'boolean: if set and not empty, stream wrapper is disabled, otherwise the stream wrapper is enabled and streams are only fetched when reading from for the first time. If your code always uses all binary properties it reads, you can disable this for a small performance gain.',
         'jackalope.logger' => 'Psr\Log\LoggerInterface: Use the LoggingClient to wrap the default transport Client',
+        Session::OPTION_AUTO_LASTMODIFIED => 'boolean: Whether to automatically update nodes having mix:lastModified. Defaults to true.',
     );
 
     /**
@@ -97,6 +98,9 @@ class RepositoryFactoryJackrabbit implements RepositoryFactoryInterface
         }
 
         $options['stream_wrapper'] = empty($parameters['jackalope.disable_stream_wrapper']);
+        if (isset($parameters[Session::OPTION_AUTO_LASTMODIFIED])) {
+            $options[Session::OPTION_AUTO_LASTMODIFIED] = $parameters[Session::OPTION_AUTO_LASTMODIFIED];
+        }
 
         return new Repository($factory, $transport, $options);
     }