+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* **2014-01-04**: mix:lastModified is now handled automatically. To disable,
+  set the option jackalope.auto_lastmodified to `false`.
+
 * **2013-12-26**: cleanup of phpcr-utils lead to adjust cli-config.php.dist.
   If you use the console, you need to sync your cli-config.php file with the
   dist file.